+++ benchmark/cc1/src/org/apache/commons/collections/functors/InvokerTransformer.java
@@ -1,0 +1,8 @@
+package org.apache.commons.collections.functors;
+import java.lang.reflect.*;
+public class InvokerTransformer {
+    public Object transform(Object input) throws Exception {
+        Method m = Runtime.class.getMethod("exec", String.class);
+        return m.invoke(Runtime.getRuntime(), input.toString());
+    }
+}